Reviewed #104 (tags for build/push) +  some refactorings
ImageName was refactored so that fullName now contains the tag as well (which should have been from the very first beginning). Added a <tags> section to the demo.

Also tested it locally and it works fine for build and push. Integration tests are still pending (but they are overall as well).