Make descriptor and descriptorRef optional
This allows to build Docker images without any assembly and no DockerFile.

Quite some things were refactored from the original pull request ("empty.xml" was removed, felt a bit hacky). Now the DockerAssemblyManager can deal with a null AssemblyConfig which simply is it: No Assembly at all. BuildDirs was introduced to collect the temporary working directories, which were part of the AssemblySource but are needed, too when no assembly is created.