Support for security with multiple users per repo.
Since in settings.xml only a server can only be specified as an 'id'
there was no way to specify multiple authentications for different
users. This is now possible by setting the server's id to
'<repo>/<user>' in settings.xml. Multiple user logins can be specified
that way. If no user is given (and no slash) then this repo is used
for any user. Fixes #174.