Http client now uses response handlers to avoid leaving open connections.
The http client was not correctly used which was leaving connections open and not being able to reuse, which caused at some point to reach the max connections per route. We faced this particular issue when trying to build 2 images with 2 tags each: since the default limit of connections per route host is 2 when we tried to build the images the plugin was getting stuck trying to get more connections.
The change consists mainly in changing execute with returns of result with proper ResponseHandlers so http client library automatically handles response automatic close to release open connections.

Signed-off-by: Roger Abelenda <rabelenda@gmail.com>