Allow ${...} for portmappings, too.
This works nicely if you don't want to write out a property file with port and host values. Otherwise you need to use the old syntax since when using ${..} the variable name is not preserved because when the property is set then it gets interpolated by maven itself before the plugin gets it. Only when its empty, the property name is injected in which case the plugin can fill it with the dynamically assigned port.

tl;dr : Use ${...} if you don't need to write out port properties in a file.

In a future version we will provide a property mode with writes out all properties with a calculated name so that there is no need to preserve property names.

See also #240 for a lengthy discussion on that.