Don't use image username for auto pull
If a default login for the registry is configured for authentication this will be used, regardless of the image's username. If no default login is registered, no authentication is used.

This affects only pulling. For push, the default user is extracted from the image name (but can be overwritten with options and config).

This also will fix fabric8io/fabric8-devops#46

Relates to #147