Refactor ServiceHub
In order to avoid dangerous and tedious state management for the initialization of ServiceHub, the ServiceHub is now created for each run afresh from a ServiceHubFactory which is needed because of collection Maven dependencies (hence it is configured as a component).

This will allow multiple docker host configurations again and will fix #320.

Some minor cleanup is probably still needed.

ServiceHub should probably called DmpContext as it is now the context given through from which all backend services (which are stateful because of the DockerUrl) are reachable.