#350: Implemented support for OpenShift authentication
Now, if either `-Ddocker.useOpenShiftAuth` or `<useOpenShiftAuth>true</useOpenShiftAuth>` in the plugin configuration is used, then the currently connected OpenShift user is used for authentication in pull and/or push operation. This is typicallt performed with a 'oc login'. Using this features correspond to use `oc whoami`  for the user part and `oc whoami -t` for the token which is to be used as password.