#351: Support for dedicarted pull and push registries added.
Registries can now be specified with 'docker.pull.registry' and 'docker.push.registry' (with a fallback 'docker.registry') as properties (or '<pullRegistry>' and '<pushRegistry>' in the plugin configuration.
Authentication can be specified also separately with 'docker.pull.username' / 'docker.push.username', ... and within the configuration in '<authConfig><push><username>...</username></push></authConfig' ....

Also added support autopulling and authentication of base images included in dockerfiles when 'dockerFileDir' is used. This fixes #358.