#401 : Extended cleanup with 'try'
The new defaul 'try' will try to remove an image but wont fail if not possible (e.g. because its still in use by a running container). Alternative values are 'none' and 'remove' (which fails when removing fails).
For backwards compatibility, also 'false' and 'true' are supported (but changed to reflect that its not a binary value anymore).

Fixes #401.