Update readStreamFrame() to error when a partial buffer is read.
When zero bytes are read when trying to read the 8 byte header, assume the
stream is done and stop iterating. If 1-7 bytes are read when trying to read
the 8 byte header, throw an IOException indicating the error. If anything other
than the full N bytes are read for the message, i.e. if 0 to N-1 bytes are
read, also throw an IOException.

This ensures that unexpected log stream behavior is propaged as an error to the
user.

Signed-off-by: Thomas Van Doren <thomas.vandoren@gmail.com>