#430 : Add a new <mode> for <tcp> wait configuration
Can bei either "mapped" which uses a remote Docker IP and mapped ports or "direct" where the container is contacted directly with the unmapped ports (needed to avoid issues with the user-proxy of Docker daemons which are enabled by default).

"mapped" is the default when host is not "localhost", "direct" when host is "localhost".