#384 : Rebased to integration branch.
Was a bit of a work because package has been renamed in the meantime to io.fabric8.
Also started to restructure the config handling to be more flexible and extensible for new config handlers.

The test are not yet running because the juse mockito which has been removed in the meantime. We are using jmockit as unique Mocking framework now.

Story to be continued ...