#513 : Changed getByName() to no use listContainers()
Apparently this works for every supported version down to 1.18 (Docker 1.6), since `/containers/<id or name>/json` can use both id and name. We used it only for ids so fart, but this is changed with this commit. Also there is now DockerAccess.getServerApiCVersion() to detect the API version running on the server.