Removed build-nofork and source-nofork
Instead the missing artifact file (which is set by the 'package' packager and hence must run before docker:build) is calculated by the DockerAssemblyManager itself.

This is not bullet proof (e.g. doesnt work when using classifier for the jar) but should work in the majority of use cases.

The error message has been improved for the case when this still not works. As before one can avoid the problem by calling 'mvn package docker:build'.

BTW, the reason for this (non-changeable) Maven behaviour is described in https://issues.apache.org/jira/browse/MASSEMBLY-94