Add new assembly option "name"
this can be used to specify the assembly name which is used also as directory
for holding the assembly files. I.e. the directory which needs to be refered to when
an extrean Dockerfile wants to include an assembly. "targetDir" now defaults to "/" + this name.

Also extended the documentation to clarify the usage.

Fixes #634