Don't fail build if volume-clean fails.
This fix is admittedly simplistic.  It does not take into account any of the
other possible causes for an exception.  However, if you are using volume-clean,
chances are you're using volume-create, or many of the other goals provided by
DMP.  In that case, you will almost certainly receive errors pointing you at the
real problem.