Remove infected test image
This image archive (included since version 0.20.1) is used in an integration test only and is removed immediately after the test again. The image was also never executed and only used for testing the image save / load feature of d-m-p.

It is *not* part of any released Maven artefact. So I consider the impact quite low.

Fixes #795