Add option for regenerating Docker Machine certificates
If using Docker Machine with the AWS EC2 driver, the remote host will
have a new IP address every time it is started up again.  This causes
the Docker Maven plug-in to fail with an error about certificates
being invalid when starting up the machine automatically.  Add an
option to automatically regenerate the certificates for the new IP
address after startup.  Setting this option to true will allow Docker
Machines running in AWS EC2 to be automatically started up into a
usable state by the Docker Maven plug-in.

Signed-off-by: Markus Hallmann <markhalm256@gmail.com>