Fix: Smarter API version handling (#1095)
Fixes #1060 Send a test request to the Daemon, and check whether it succeeds. In the return response
there's a header Api-Version: which holds the actual Docker daemon version. If this test
request because of a too low version, we can just pick up that version and use it as
minimal version for us.