Use top-level authentication for extended auth
When relying on [Extended Authentication](https://dmp.fabric8.io/#extended-authentication) for ECR, we unfortunately have to configure the `push` and `pull` sections, even though they are the same. The problem seems to be that the AWS session token has to be put into the `<auth>` element, whereas the general section only knows an `<authToken>` element. This change would allow Extended Authentication to work with the general configuration.

When using system properties on the other hand, the correct property is `docker.auth` instead. So maybe the proper way forward would be to deprecate the `authToken` property and introduce `auth` - or vice versa.