fix : `FROM --platform` in Dockerfile causes build error (#1516)
Right now we need to pull base images (ones specified in `FROM`
statements) before doing the actual build. DockerfileUtil doesn't seem
to be handling the case when an option is specified in `FROM` statement.
It just assumes second token to be image name.

Add check to ignore token starting with `--` so that it doesn't fail the
build. For specifiying pull platform it would need to be set in image
configuration build createImageOptions

Signed-off-by: Rohan Kumar <rohaan@redhat.com>