feat(config): extend HealthCheckMode with new shell mode
Docker Compose and Docker API distinguish between cmd (array of args) and cmd-shell (command string) mode, so we need to add this as a mode for our goal to enable runtime defined healthchecks.
Users will need to pay attention which mode they use in runtime defined checks, as the Dockerfile buildtime check is not making this difference!