feat(config): add runtime healthcheck mode inherit
In case a user provides a healthcheck specification as part of runtime configuration or compose, the intent might be to alter the options only and no the test itself.

An explicit mode that needs to be added to the XML config or determined by the Compose handler makes users' intentions clearer and easier to validate.

Also merging shell case in DockerFileBuilder to default as we have two invalid nodes now.